Date: February 2, 2016 at 11:51:58 AM EST
Removing dispatch draw. That's actually called to draw children of a view. so the arc is technically drawn before the circle, but it all happens too quickly for the eye to catch.
I need to fix butterknife in the view class, it's not binding the views because I'm setting it up in the wrong place