Date: February 9, 2016 at 1:11:05 PM EST
Returning to development on Study Pro.
I'm going to tackle the notes page since I feel I can do it in a few hours and then return to the timer. I just saw the database class I wrote back in college. It's pretty bad... I'm scrapping it for Ormlite instead.